rename manager to wait_children
